requirements.txt
# Core
networkx>=3.0
numpy>=1.24.0
pillow>=10.0.0
PyYAML>=6.0
requests>=2.31.0
python-dotenv>=1.0.0
matplotlib>=3.7.0

# NLP (ambiguity detector)
torch
transformers
spacy
spacy-transformers
pandas
scikit-learn
nltk
tqdm

# API
openai>=1.0.0

# Web server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
